+++ app/Http/Controllers/API/TableController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Restaurant;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Http\Resources\TableResource;
+use Illuminate\Support\Facades\Gate;
 
 class TableController extends Controller
 {
@@ -18,29 +19,38 @@ class TableController extends Controller
 
     public function store(Request $request, Restaurant $restaurant)
     {
-        $table = $restaurant->tables()->create($request->all());
-        return new TableResource($table);
-    }
-
-    public function update(Request $request, Restaurant $restaurant, $id)
-    {
-        try {
-            $table = $restaurant->tables()->findOrFail($id);
-            $table->update($request->all());
+        if(Gate::allows('manage', $restaurant)) {
+            $table = $restaurant->tables()->create($request->all());
             return new TableResource($table);
-        } catch(ModelNotFoundException $exception) {
-            return response()->json(['error' => $exception->getMessage()], 404);
         }
+        return response()->json(['message' => 'Unauthorized'], 401);
     }
 
-    public function delete(Restaurant $restaurant, $id)
+    public function update(Request $request, Restaurant $restaurant, $table)
     {
-        try {
-            $table = $restaurant->tables()->findOrFail($id);
-            $table->delete();
-            return response()->json(['message' => 'table deleted']);
-        } catch(ModelNotFoundException $exception) {
-            return response()->json(['error' => $exception->getMessage()], 404);
+        if(Gate::allows('manage', $restaurant)) {
+            try {
+                $table = $restaurant->tables()->findOrFail($table);
+                $table->update($request->all());
+                return new TableResource($table);
+            } catch(ModelNotFoundException $exception) {
+                return response()->json(['error' => $exception->getMessage()], 404);
+            }
         }
+        return response()->json(['message' => 'Unauthorized'], 401);
+    }
+
+    public function delete(Restaurant $restaurant, $table)
+    {
+        if(Gate::allows('manage', $restaurant)) {
+            try {
+                $table = $restaurant->tables()->findOrFail($table);
+                $table->delete();
+                return response()->json(['message' => 'table deleted']);
+            } catch(ModelNotFoundException $exception) {
+                return response()->json(['error' => $exception->getMessage()], 404);
+            }
+        }
+        return response()->json(['message' => 'Unauthorized'], 401);
     }
 }
